# openapi_v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EZ-Lang-Auth",
    "version": "1.0.0",
    "description": "REST API designed to provide functions for ez language",
    "contact": {
      "email": "vinhphuc.it@gmail.com",
      "name": "Phuc Tran"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/ezlang/"
    },
    {
      "url":"https://ez-lang-alpha.herokuapp.com/ezlang/"
    }
  ],
...
  "paths": {
...
    "/auth/refreshToken": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "RefreshToken",
        "operationId": "refreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfull Operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/verifyEmail":{
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "verifyEmail",
        "operationId": "verifyEmail",      
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Operation"            
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/resetPassword":{
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "resetPassword",
        "operationId": "resetPassword",      
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Operation"            
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/confirmResetPassword":{
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "confirmResetPassword",
        "operationId": "confirmResetPassword",      
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Operation"            
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/changePassword":{
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "changePassword",
        "operationId": "changePassword",      
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Operation"            
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SignInResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "number"
          },
          "refreshToken": {
            "type": "string"
          },
          "refreshTokenExpiresAt": {
            "type": "number"
          }
        }
      },
...
      "VerifyEmailRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "ConfirmResetPasswordRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "ChangePasswordRequest":{
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "oldPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}